# 성능 테스트 개선

## 변경사항
성능 테스트의 신뢰도와 안정성을 향상시키기 위한 테스트 환경 개선을 진행했습니다.

### 1. 워밍업 프로세스 최적화
```typescript
// realistic-performance.spec.ts
beforeAll(async () => {
  const WARMUP_ITERATIONS = 5;
  const WARMUP_COMPONENTS = 10;
  
  for (let i = 0; i < WARMUP_ITERATIONS; i++) {
    const warmupCode = generateRealisticProject({
      componentCount: WARMUP_COMPONENTS * (i + 1), // 점진적 증가
      maxEffectsPerComponent: 3,
      maxDependenciesPerEffect: 2,
      variant: { type: 'function', hasChildren: true, hasProps: true, hasState: true },
    });
  }
});
```

- JIT 컴파일러 워밍업 횟수 증가 (3 → 5)
- 작은 크기부터 점진적으로 증가하는 워밍업 패턴 적용
- 시스템 안정화를 위한 대기 시간 유지

### 2. 테스트 케이스 구조화
```typescript
const testCases = [
  { componentCount: 10, maxEffectsPerComponent: 3, maxDependenciesPerEffect: 2, variant: 'function' },
  { componentCount: 25, maxEffectsPerComponent: 4, maxDependenciesPerEffect: 3, variant: 'arrow' },
  { componentCount: 50, maxEffectsPerComponent: 5, maxDependenciesPerEffect: 3, variant: 'arrow' },
  { componentCount: 75, maxEffectsPerComponent: 4, maxDependenciesPerEffect: 4, variant: 'memo' },
  { componentCount: 100, maxEffectsPerComponent: 3, maxDependenciesPerEffect: 4, variant: 'memo' }
];
```

- 다양한 컴포넌트 타입과 크기에 대한 체계적인 테스트
- 실제 사용 사례를 반영한 테스트 구성

## 테스트 결과

### 1. 변환 시간 안정성
- 50개 컴포넌트: 변동폭 ~6% (3.33ms ~ 3.49ms)
- 100개 컴포넌트: 변동폭 ~10% (2.97ms ~ 3.95ms)

### 2. 메모리 사용량 일관성
- 전체 변동폭: 2-8%
- 대규모 컴포넌트(50+): 변동폭 2-3%

### 3. 번들 크기 안정성
- 번들 크기: 2.93KB ~ 3.38KB
- Gzip 크기: 0.97KB ~ 1.06KB
- 변동폭: ~13%

## 개선 효과
1. 테스트 측정의 일관성 향상
2. 대규모 컴포넌트 처리의 안정적인 성능 측정
3. 메모리 사용량 측정의 신뢰도 개선

## 테스트 실행 방법
```bash
pnpm test:performance src/__tests__/realistic-performance.spec.ts
```

## 관련 파일
- src/__tests__/realistic-performance.spec.ts
- src/__tests__/performance-utils.ts
- src/__tests__/performance-metrics.ts

## 추가 고려사항
- CI/CD 환경에서의 안정적인 테스트 실행 보장
- 향후 더 다양한 컴포넌트 패턴에 대한 테스트 확장 가능

이 PR은 테스트 코드의 신뢰성을 향상시키는 것에 초점을 맞추었으며, 실제 플러그인의 성능 측정을 더 정확하게 할 수 있도록 개선했습니다. 특히 JIT 컴파일러의 영향을 최소화하고 일관된 측정이 가능하도록 워밍업 프로세스를 최적화했습니다.

<img width="779" alt="스크린샷 2025-02-20 오전 11 28 33" src="https://github.com/user-attachments/assets/1bde4a93-884e-43f4-98eb-41ff1c79fc1b" />
<img width="776" alt="스크린샷 2025-02-20 오전 11 26 42" src="https://github.com/user-attachments/assets/653faecf-6219-466d-bef4-a555b84131f4" />
<img width="786" alt="스크린샷 2025-02-20 오전 11 22 28" src="https://github.com/user-attachments/assets/23e0a90c-c4d5-4a23-9718-2128edbf1af5" />
<img width="783" alt="스크린샷 2025-02-20 오전 11 20 57" src="https://github.com/user-attachments/assets/ab184c02-52b2-4728-8708-01e541c83a85" />
<img width="784" alt="스크린샷 2025-02-20 오전 11 19 31" src="https://github.com/user-attachments/assets/1e17ddd8-a550-4311-9523-c3995e838a6f" />